fix(codegen): use double for integer response fields to handle decimal API values
Response fields like priceWithSellerFee return decimals despite schema
saying integer. Keep long for request params where integer is correct.

+++ codegen/Lolzteam.Codegen/Emitter.cs
@@ -425,7 +425,7 @@ internal static partial class Emitter
             return type switch
             {
                 "string" => "string",
-                "integer" => "long",
+                "integer" => "double",
                 "number" => "double",
                 "boolean" => "bool",
                 _ => "JsonElement",
